+++ pkg/server/BUILD.bazel
@@ -173,7 +173,6 @@ go_test(
         "//pkg/server/internal/parse",
         "//pkg/server/internal/testutil",
         "//pkg/server/internal/util",
-        "//pkg/server/tests/servertestkit",
         "//pkg/session",
         "//pkg/sessionctx/variable",
         "//pkg/sessiontxn",
@@ -205,8 +204,6 @@ go_test(
         "@com_github_tikv_client_go_v2//oracle",
         "@com_github_tikv_client_go_v2//testutils",
         "@com_github_tikv_client_go_v2//tikv",
-        "@io_gorm_driver_mysql//:mysql",
-        "@io_gorm_gorm//:gorm",
         "@org_uber_go_goleak//:goleak",
     ],
 )
