+++ pkg/server/tests/commontest/BUILD.bazel
@@ -45,6 +45,8 @@ go_test(
         "@com_github_stretchr_testify//require",
         "@com_github_tikv_client_go_v2//tikv",
         "@com_github_tikv_client_go_v2//tikvrpc",
+        "@io_gorm_driver_mysql//:mysql",
+        "@io_gorm_gorm//:gorm",
         "@io_opencensus_go//stats/view",
         "@org_uber_go_goleak//:goleak",
     ],
