+++ pkg/server/tidb_test.go
@@ -18,206 +18,18 @@ import (
 	"bufio"
 	"bytes"
 	"context"
-	"database/sql"
 	"fmt"
-	"reflect"
 	"testing"
 
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/server/internal"
-	"github.com/pingcap/tidb/pkg/server/tests/servertestkit"
 	"github.com/pingcap/tidb/pkg/sessiontxn"
 	"github.com/pingcap/tidb/pkg/testkit"
 	"github.com/pingcap/tidb/pkg/util/arena"
 	"github.com/pingcap/tidb/pkg/util/chunk"
 	"github.com/stretchr/testify/require"
-	gorm_mysql "gorm.io/driver/mysql"
-	"gorm.io/gorm"
 )
-
-func TestVectorTypeTextProtocol(t *testing.T) {
-	// Text protocol is used in non-prepared query (COM_QUERY).
-	// See https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_com_query_response_text_resultset.html
-	ts := servertestkit.CreateTidbTestSuite(t)
-	ts.RunTests(t, nil, func(dbt *testkit.DBTestKit) {
-		dbt.MustExec("DROP TABLE IF EXISTS test")
-		dbt.MustExec("CREATE TABLE test (a VECTOR, b VECTOR(3))")
-		dbt.MustExec("INSERT INTO test VALUES ('[]', '[1,2,3]')")
-
-		rows := dbt.MustQuery("SELECT * FROM test")
-
-		// Check column types
-		columnTypes, err := rows.ColumnTypes()
-		require.NoError(t, err)
-		require.Len(t, columnTypes, 2)
-		require.Equal(t, "LONGTEXT", columnTypes[0].DatabaseTypeName())
-		// https://github.com/go-sql-driver/mysql/blob/v1.7.1/fields.go#L195
-		require.Equal(t, reflect.TypeOf(sql.RawBytes{}).String(), columnTypes[0].ScanType().String())
-		require.Equal(t, "LONGTEXT", columnTypes[1].DatabaseTypeName())
-		require.Equal(t, reflect.TypeOf(sql.RawBytes{}).String(), columnTypes[1].ScanType().String())
-
-		require.True(t, rows.Next())
-		rowDatums := make([]any, 2)
-		rowDatumsPtr := make([]any, 2)
-		for i := range rowDatumsPtr {
-			rowDatumsPtr[i] = &rowDatums[i]
-		}
-
-		err = rows.Scan(rowDatumsPtr...)
-		require.NoError(t, err)
-
-		require.Equal(t, []byte("[]"), rowDatums[0])
-		require.Equal(t, []byte("[1,2,3]"), rowDatums[1])
-
-		require.False(t, rows.Next())
-		require.NoError(t, rows.Close())
-
-		// When using string, the driver will return the data as string.
-		rows = dbt.MustQuery("SELECT * FROM test")
-		require.NoError(t, err)
-
-		require.True(t, rows.Next())
-		var valA, valB string
-		err = rows.Scan(&valA, &valB)
-		require.NoError(t, err)
-
-		require.Equal(t, "[]", valA)
-		require.Equal(t, "[1,2,3]", valB)
-
-		require.False(t, rows.Next())
-		require.NoError(t, rows.Close())
-
-		// Also work with VECTOR immediate values.
-		rows = dbt.MustQuery("SELECT VEC_FROM_TEXT('[1,2]')")
-		require.NoError(t, err)
-
-		columnTypes, err = rows.ColumnTypes()
-		require.NoError(t, err)
-		require.Len(t, columnTypes, 1)
-		require.Equal(t, "LONGTEXT", columnTypes[0].DatabaseTypeName())
-		require.Equal(t, reflect.TypeOf(sql.RawBytes{}), columnTypes[0].ScanType())
-
-		require.True(t, rows.Next())
-		err = rows.Scan(&valA)
-		require.NoError(t, err)
-
-		require.Equal(t, "[1,2]", valA)
-
-		require.False(t, rows.Next())
-		require.NoError(t, rows.Close())
-	})
-}
-
-func TestVectorTypeBinaryProtocol(t *testing.T) {
-	// Binary protocol is used in prepared statements (COM_STMT_EXECUTE).
-	// See https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_binary_resultset.html
-	ts := servertestkit.CreateTidbTestSuite(t)
-	ts.RunTests(t, nil, func(dbt *testkit.DBTestKit) {
-		dbt.MustExec("DROP TABLE IF EXISTS test")
-		dbt.MustExec("CREATE TABLE test (a VECTOR, b VECTOR(3))")
-		dbt.MustExec("INSERT INTO test VALUES ('[]', '[1,2,3]')")
-
-		stmt := dbt.MustPrepare("SELECT * FROM test")
-		defer stmt.Close()
-
-		// When using interface{}, the driver will return the data as []byte.
-		// Note: This is the same behavior as TEXT type.
-		rows, err := stmt.Query()
-		require.NoError(t, err)
-
-		columnTypes, err := rows.ColumnTypes()
-		require.NoError(t, err)
-		require.Len(t, columnTypes, 2)
-		require.Equal(t, "LONGTEXT", columnTypes[0].DatabaseTypeName())
-		// https://github.com/go-sql-driver/mysql/blob/v1.7.1/fields.go#L195
-		require.Equal(t, reflect.TypeOf(sql.RawBytes{}).String(), columnTypes[0].ScanType().String())
-		require.Equal(t, "LONGTEXT", columnTypes[1].DatabaseTypeName())
-		require.Equal(t, reflect.TypeOf(sql.RawBytes{}).String(), columnTypes[1].ScanType().String())
-
-		require.True(t, rows.Next())
-		rowDatums := make([]any, 2)
-		rowDatumsPtr := make([]any, 2)
-		for i := range rowDatumsPtr {
-			rowDatumsPtr[i] = &rowDatums[i]
-		}
-
-		err = rows.Scan(rowDatumsPtr...)
-		require.NoError(t, err)
-
-		require.Equal(t, []byte("[]"), rowDatums[0])
-		require.Equal(t, []byte("[1,2,3]"), rowDatums[1])
-
-		require.False(t, rows.Next())
-		require.NoError(t, rows.Close())
-
-		// When using string, the driver will return the data as string.
-		rows, err = stmt.Query()
-		require.NoError(t, err)
-
-		require.True(t, rows.Next())
-		var valA, valB string
-		err = rows.Scan(&valA, &valB)
-		require.NoError(t, err)
-
-		require.Equal(t, "[]", valA)
-		require.Equal(t, "[1,2,3]", valB)
-
-		require.False(t, rows.Next())
-		require.NoError(t, rows.Close())
-
-		// Also work with VECTOR immediate values.
-		stmt2 := dbt.MustPrepare("SELECT VEC_FROM_TEXT('[1,2]')")
-		defer stmt2.Close()
-
-		rows, err = stmt2.Query()
-		require.NoError(t, err)
-
-		columnTypes, err = rows.ColumnTypes()
-		require.NoError(t, err)
-		require.Len(t, columnTypes, 1)
-		require.Equal(t, "LONGTEXT", columnTypes[0].DatabaseTypeName())
-		require.Equal(t, reflect.TypeOf(sql.RawBytes{}), columnTypes[0].ScanType())
-
-		require.True(t, rows.Next())
-		err = rows.Scan(&valA)
-		require.NoError(t, err)
-
-		require.Equal(t, "[1,2]", valA)
-
-		require.False(t, rows.Next())
-		require.NoError(t, rows.Close())
-	})
-}
-
-type VectorTestModel struct {
-	ID         int
-	Embedding  string `gorm:"type:vector(3)"`
-	Embedding2 string `gorm:"type:vector(3)"`
-}
-
-func TestVectorTypeGORM(t *testing.T) {
-	ts := servertestkit.CreateTidbTestSuite(t)
-	ts.RunTestsOnNewDB(t, nil, "vector_db", func(dbt *testkit.DBTestKit) {
-		dbgorm, err := gorm.Open(gorm_mysql.New(gorm_mysql.Config{
-			Conn: dbt.GetDB(),
-		}), &gorm.Config{})
-		require.NoError(t, err)
-
-		require.NoError(t, dbgorm.AutoMigrate(&VectorTestModel{}))
-
-		tx := dbgorm.Create(&VectorTestModel{ID: 10, Embedding: "[1,2.0,3.0]", Embedding2: "[2,2,2]"})
-		require.NoError(t, tx.Error)
-
-		var v VectorTestModel
-		tx = dbgorm.First(&v, "id = ?", 10)
-		require.NoError(t, tx.Error)
-
-		require.Equal(t, 10, v.ID)
-		require.Equal(t, "[1,2,3]", v.Embedding)
-		require.Equal(t, "[2,2,2]", v.Embedding2)
-	})
-}
 
 func TestRcReadCheckTSConflict(t *testing.T) {
 	store := testkit.CreateMockStore(t)
